@abstractdog 2022-01-30
double-check full unit test
before: [INFO] Total time:  10:53 min (Wall Clock)
after: [INFO] Total time:  03:47 min (Wall Clock)

merging this, thanks @mbathori-cloudera  for the patch and testing